requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
firebase-admin==5.0.0
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
httplib2==0.19.1
idna==2.10
mdutils==1.3.0
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.26.4
fpdf==1.7.2